mirage-os-shim provides only an interface
The goal of this PR is to provide a proper way to have
an abstraction of entry-points of underlying target of a MirageOS.

According to the target, it shoud provide some ways to register
some function at the beginning and at the end of the MirageOS process.
mirage-os-shim should be an only interface to describe that.